+++ drivers/sqlite3/sqlite3.go
@@ -15,7 +15,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/neilotoole/sq/libsq/core/dialect"
+	"github.com/neilotoole/sq/libsq/driver/dialect"
 
 	"github.com/neilotoole/sq/libsq/core/lg/lga"
